+++ History.md
@@ -17,6 +17,7 @@
 - Bugfix: Some changes to the proposed procedure created a new version of it, although it did not make sense (e.g. publishing it).
 - Bugfix: Automatic scrolling to motion/amendment comments was not working reliably.
 - Bugfix: If a modified adoption of an amendment was proposed, and the amendment was later on deleted, then merging other amendments into the base motion failed.
+- Bugfix: The All-in-one PDF export of motions and amendments for users took into account the filter from the admin motion list, if set.
 
 ### Version 4.17.1 (2026-06-22)